+++ chaos/common/thread/WaitSemaphore.h
@@ -86,7 +86,7 @@ namespace chaos {
          if no one unlock it first
          \return the object that someone has associated at this semaphore
          */
-        void wait(unsigned int millisecToWait) {
+        void wait(unsigned long millisecToWait) {
             boost::unique_lock<boost::mutex> lock( wait_answer_mutex );
             if(inWait) return;
             inWait = true;
